requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
pymongo==4.5.0
python-dotenv==1.0.0
Werkzeug==2.3.7
email-validator==2.0.0
Flask-SocketIO==5.3.4
python-socketio==5.8.0
eventlet==0.33.3 